+++ db/migrate/20100118174556_add_render_link.rb
@@ -1,9 +1,0 @@
-class AddRenderLink < ActiveRecord::Migration
-  def self.up
-    add_column :repositories, :render_link, :boolean, :null => true
-  end
-
-  def self.down
-    remove_column :repositories, :render_link
-  end
-end