+++ lib/open_project/checkout/engine.rb
@@ -60,6 +60,14 @@ Please select the desired protocol below to get the URL.
       FactoryGirl.definition_file_paths << File.expand_path(self.root.to_s + '/spec/factories') if defined?(FactoryGirl)
     end
 
+    initializer :append_migrations do |app|
+      unless app.root.to_s.match root.to_s
+        config.paths["db/migrate"].expanded.each do |expanded_path|
+          app.config.paths["db/migrate"] << expanded_path
+        end
+      end
+    end
+
     spec = Bundler.environment.specs['openproject-checkout'][0]
     initializer "checkout.register_plugin" do
       require_dependency 'open_project/checkout/patches/repository_patch'